==> 1_1_DataProcessing.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
ad_feature = pd.read_csv('/content/drive/MyDrive/Taobao/ad_feature.csv')
raw_sample = pd.read_csv('/content/drive/MyDrive/Taobao/raw_sample.csv')
user_profile = pd.read_csv('/content/drive/MyDrive/Taobao/user_profile.csv')
# behavior_log = pd.read_csv('/content/drive/MyDrive/Taobao/behavior_log.csv')

raw_sample = raw_sample.rename(columns={'user':'userid'})

all_data = pd.merge(pd.merge(raw_sample,ad_feature),user_profile)
all_data['time'] = pd.to_datetime(all_data['time_stamp'], unit='s') + pd.Timedelta('08:00:00')

all_data['day'] = all_data['time'].dt.day
all_data['hour'] = all_data['time'].dt.hour

all_data.to_csv('/content/drive/MyDrive/Taobao/all_data.csv')
