==> 1_3_FeatureExtraction.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import numpy as np
import pandas as pd
import math
import gc
import os

data = pd.read_csv('/content/drive/MyDrive/Taobao/all_data.csv')

data['day_hour'] = data['day']*24 + data['hour']
data['weekday'] = data['day']%7

data.rename(columns = {'new_user_class_level ':'new_user_class_level'},inplace = True)

data.columns

def get_results(original_df, group_idx_list, offset_idx, offset, new_column, aggregate_method = 'count'):
  if (aggregate_method == "count"):
    temp_df = original_df.groupby(group_idx_list)['clk'].count().reset_index()
  elif (aggregate_method == "sum"):
    temp_df = original_df.groupby(group_idx_list)['clk'].sum().reset_index()
  else:
    print('Error')
    return

  temp_df[offset_idx] += offset
  temp_df.rename(columns = {'clk':new_column}, inplace = True)

  return temp_df

#### code

userid + adgroupid

# 过去一小时的展示次数
temp_num_show_last_hour = get_results(data, ['userid','adgroup_id','day_hour'], 'day_hour', 1, 'num_show_last_hour_user_adgroup','count')
# 过去一小时的点击次数
temp_num_click_last_hour = get_results(data, ['userid','adgroup_id','day_hour'], 'day_hour', 1, 'num_click_last_hour_user_adgroup','sum')
# 过去一天的展示次数
temp_num_show_last_day = get_results(data, ['userid','adgroup_id','day'], 'day', 1, 'num_show_last_day_user_adgroup','count')
# 过去一天的点击次数
temp_num_click_last_day = get_results(data, ['userid','adgroup_id','day'], 'day', 1, 'num_click_last_day_user_adgroup','sum')

data = pd.merge(data,temp_num_show_last_hour,how="left")
del temp_num_show_last_hour
gc.collect()

data = pd.merge(data,temp_num_click_last_hour,how="left")
del temp_num_click_last_hour
gc.collect()

data = pd.merge(data,temp_num_show_last_day,how="left")
del temp_num_show_last_day
gc.collect()

data = pd.merge(data,temp_num_click_last_day,how="left")
del temp_num_click_last_day
gc.collect()

userid + cate_id

# 过去一小时的展示次数
temp_num_show_last_hour = get_results(data, ['userid','cate_id','day_hour'], 'day_hour', 1, 'num_show_last_hour_user_cate','count')
# 过去一小时的点击次数
temp_num_click_last_hour = get_results(data, ['userid','cate_id','day_hour'], 'day_hour', 1, 'num_click_last_hour_user_cate','sum')
# 过去一天的展示次数
temp_num_show_last_day = get_results(data, ['userid','cate_id','day'], 'day', 1, 'num_show_last_day_user_cate','count')
# 过去一天的点击次数
temp_num_click_last_day = get_results(data, ['userid','cate_id','day'], 'day', 1, 'num_click_last_day_user_cate','sum')

data = pd.merge(data,temp_num_show_last_hour,how="left")
del temp_num_show_last_hour
gc.collect()

data = pd.merge(data,temp_num_click_last_hour,how="left")
del temp_num_click_last_hour
gc.collect()

data = pd.merge(data,temp_num_show_last_day,how="left")
del temp_num_show_last_day
gc.collect()

data = pd.merge(data,temp_num_click_last_day,how="left")
del temp_num_click_last_day
gc.collect()

userid + customer_id

# 过去一小时的展示次数
temp_num_show_last_hour = get_results(data, ['userid','customer','day_hour'], 'day_hour', 1, 'num_show_last_hour_user_customer','count')
# 过去一小时的点击次数
temp_num_click_last_hour = get_results(data, ['userid','customer','day_hour'], 'day_hour', 1, 'num_click_last_hour_user_customer','sum')
# 过去一天的展示次数
temp_num_show_last_day = get_results(data, ['userid','customer','day'], 'day', 1, 'num_show_last_day_user_customer','count')
# 过去一天的点击次数
temp_num_click_last_day = get_results(data, ['userid','customer','day'], 'day', 1, 'num_click_last_day_user_customer','sum')

data = pd.merge(data,temp_num_show_last_hour,how="left")
del temp_num_show_last_hour
gc.collect()

data = pd.merge(data,temp_num_click_last_hour,how="left")
del temp_num_click_last_hour
gc.collect()

data = pd.merge(data,temp_num_show_last_day,how="left")
del temp_num_show_last_day
gc.collect()

data = pd.merge(data,temp_num_click_last_day,how="left")
del temp_num_click_last_day
gc.collect()

userid + brand

# 过去一小时的展示次数
temp_num_show_last_hour = get_results(data, ['userid','brand','day_hour'], 'day_hour', 1, 'num_show_last_hour_user_brand','count')
# 过去一小时的点击次数
temp_num_click_last_hour = get_results(data, ['userid','brand','day_hour'], 'day_hour', 1, 'num_click_last_hour_user_brand','sum')
# 过去一天的展示次数
temp_num_show_last_day = get_results(data, ['userid','brand','day'], 'day', 1, 'num_show_last_day_user_brand','count')
# 过去一天的点击次数
temp_num_click_last_day = get_results(data, ['userid','brand','day'], 'day', 1, 'num_click_last_day_user_brand','sum')

data = pd.merge(data,temp_num_show_last_hour,how="left")
del temp_num_show_last_hour
gc.collect()

data = pd.merge(data,temp_num_click_last_hour,how="left")
del temp_num_click_last_hour
gc.collect()

data = pd.merge(data,temp_num_show_last_day,how="left")
del temp_num_show_last_day
gc.collect()

data = pd.merge(data,temp_num_click_last_day,how="left")
del temp_num_click_last_day
gc.collect()

userid + all

# 过去一小时的展示次数
temp_num_show_last_hour = get_results(data, ['userid','day_hour'], 'day_hour', 1, 'num_show_last_hour_user_all','count')
# 过去一小时的点击次数
temp_num_click_last_hour = get_results(data, ['userid','day_hour'], 'day_hour', 1, 'num_click_last_hour_user_all','sum')
# 过去一天的展示次数
temp_num_show_last_day = get_results(data, ['userid','day'], 'day', 1, 'num_show_last_day_user_all','count')
# 过去一天的点击次数
temp_num_click_last_day = get_results(data, ['userid','day'], 'day', 1, 'num_click_last_day_user_all','sum')

data = pd.merge(data,temp_num_show_last_hour,how="left")
del temp_num_show_last_hour
gc.collect()

data = pd.merge(data,temp_num_click_last_hour,how="left")
del temp_num_click_last_hour
gc.collect()

data = pd.merge(data,temp_num_show_last_day,how="left")
del temp_num_show_last_day
gc.collect()

data = pd.merge(data,temp_num_click_last_day,how="left")
del temp_num_click_last_day
gc.collect()

all + adgroupid

# 过去一小时的展示次数
temp_num_show_last_hour = get_results(data, ['adgroup_id','day_hour'], 'day_hour', 1, 'num_show_last_hour_all_adgroup','count')
# 过去一小时的点击次数
temp_num_click_last_hour = get_results(data, ['adgroup_id','day_hour'], 'day_hour', 1, 'num_click_last_hour_all_adgroup','sum')
# # 过去一天的展示次数
# temp_num_show_last_day = get_results(data, ['adgroup_id','day'], 'day', 1, 'num_show_last_day_all_adgroup','count')
# # 过去一天的点击次数
# temp_num_click_last_day = get_results(data, ['adgroup_id','day'], 'day', 1, 'num_click_last_day_all_adgroup','sum')

data = pd.merge(data,temp_num_show_last_hour,how="left")
del temp_num_show_last_hour
gc.collect()

data = pd.merge(data,temp_num_click_last_hour,how="left")
del temp_num_click_last_hour
gc.collect()

# data = pd.merge(data,temp_num_show_last_day,how="left")
# del temp_num_show_last_day
# gc.collect()

# data = pd.merge(data,temp_num_click_last_day,how="left")
# del temp_num_click_last_day
# gc.collect()

all + cate_id

# 过去一小时的展示次数
temp_num_show_last_hour = get_results(data, ['cate_id','day_hour'], 'day_hour', 1, 'num_show_last_hour_all_cate','count')
# 过去一小时的点击次数
temp_num_click_last_hour = get_results(data, ['cate_id','day_hour'], 'day_hour', 1, 'num_click_last_hour_all_cate','sum')
# # 过去一天的展示次数
# temp_num_show_last_day = get_results(data, ['cate_id','day'], 'day', 1, 'num_show_last_day_all_cate','count')
# # 过去一天的点击次数
# temp_num_click_last_day = get_results(data, ['cate_id','day'], 'day', 1, 'num_click_last_day_all_cate','sum')

data = pd.merge(data,temp_num_show_last_hour,how="left")
del temp_num_show_last_hour
gc.collect()

data = pd.merge(data,temp_num_click_last_hour,how="left")
del temp_num_click_last_hour
gc.collect()

# data = pd.merge(data,temp_num_show_last_day,how="left")
# del temp_num_show_last_day
# gc.collect()

# data = pd.merge(data,temp_num_click_last_day,how="left")
# del temp_num_click_last_day
# gc.collect()

all + customer_id

# 过去一小时的展示次数
temp_num_show_last_hour = get_results(data, ['customer','day_hour'], 'day_hour', 1, 'num_show_last_hour_all_customer','count')
# 过去一小时的点击次数
temp_num_click_last_hour = get_results(data, ['customer','day_hour'], 'day_hour', 1, 'num_click_last_hour_all_customer','sum')
# # 过去一天的展示次数
# temp_num_show_last_day = get_results(data, ['customer','day'], 'day', 1, 'num_show_last_day_all_customer','count')
# # 过去一天的点击次数
# temp_num_click_last_day = get_results(data, ['customer','day'], 'day', 1, 'num_click_last_day_all_customer','sum')

data = pd.merge(data,temp_num_show_last_hour,how="left")
del temp_num_show_last_hour
gc.collect()

data = pd.merge(data,temp_num_click_last_hour,how="left")
del temp_num_click_last_hour
gc.collect()

# data = pd.merge(data,temp_num_show_last_day,how="left")
# del temp_num_show_last_day
# gc.collect()

# data = pd.merge(data,temp_num_click_last_day,how="left")
# del temp_num_click_last_day
# gc.collect()

all + brand

# 过去一小时的展示次数
temp_num_show_last_hour = get_results(data, ['brand','day_hour'], 'day_hour', 1, 'num_show_last_hour_all_brand','count')

data = pd.merge(data,temp_num_show_last_hour,how="left")
del temp_num_show_last_hour
gc.collect()


# 过去一小时的点击次数
temp_num_click_last_hour = get_results(data, ['brand','day_hour'], 'day_hour', 1, 'num_click_last_hour_all_brand','sum')

data = pd.merge(data,temp_num_click_last_hour,how="left")
del temp_num_click_last_hour
gc.collect()

# # 过去一天的展示次数
# temp_num_show_last_day = get_results(data, ['brand','day'], 'day', 1, 'num_show_last_day_all_brand','count')

# data = pd.merge(data,temp_num_show_last_day,how="left")
# del temp_num_show_last_day
# gc.collect()


# # 过去一天的点击次数
# temp_num_click_last_day = get_results(data, ['brand','day'], 'day', 1, 'num_click_last_day_all_brand','sum')

# data = pd.merge(data,temp_num_click_last_day,how="left")
# del temp_num_click_last_day
# gc.collect()

#### output

data

data.to_csv("./drive/MyDrive/Taobao/all_data_new.csv",index=False)
