==> 1_4_DataSampling.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import numpy as np
import pandas as pd
import math
import gc
import os

from sklearn.model_selection import train_test_split

loop = True
first = True
chunkSize = 1000000
# chunks = []

reader = pd.read_csv('/content/drive/MyDrive/Taobao/all_data_new.csv', iterator=True)

while loop:
    try:
        chunk = reader.get_chunk(chunkSize)
        if (first):
          data = chunk
          first = False
        else:
          data = pd.concat([data,chunk], ignore_index = True)
    except StopIteration:
        loop = False
        print("Iteration is stopped.")
# df = pd.concat(chunks, ignore_index=True)

import random
random.seed(123)
valid_userid = random.sample(list(set(data['userid'])), int(0.1*len(set(data['userid']))))

import gc
del data
gc.collect()

loop = True
first = True
chunkSize = 1000000

reader = pd.read_csv('/content/drive/MyDrive/Taobao/all_data_new.csv', iterator=True)

while loop:
    try:
        chunk = reader.get_chunk(chunkSize)
        chunk = chunk[chunk['userid'].isin(valid_userid)]
        if (first):
          valid_data = chunk
          first = False
        else:
          valid_data = pd.concat([valid_data,chunk], ignore_index = True)
    except StopIteration:
        loop = False
        print("Iteration is stopped.")

valid_data = valid_data[valid_data['day']>6]
valid_data = valid_data.sort_values('userid').reset_index()
valid_data.drop(columns=['Unnamed: 0', 'Unnamed: 0.1'], inplace = True)

valid_data.to_csv("./drive/MyDrive/Taobao/valid_data_10_percent.csv", index=False)
