==> 2_1_FFM.ipynb <==
!pip install xlearn

from google.colab import drive
drive.mount('/content/drive')

import xlearn as xl
import numpy as np
import pandas as pd
import math
import gc
import os

from sklearn.model_selection import train_test_split

valid_data = pd.read_csv('/content/drive/MyDrive/Taobao/valid_data_10_percent.csv')

valid_data.columns

num_train = len(valid_data[valid_data['userid']<=valid_data.loc[int(0.8*len(valid_data))]['userid']])
num_train

# numerics = ['price']  

numerics = ['price', 
      #   'num_show_last_hour_user_adgroup', 'num_click_last_hour_user_adgroup',
      #  'num_show_last_day_user_adgroup', 'num_click_last_day_user_adgroup',
      #  'num_show_last_hour_user_cate', 'num_click_last_hour_user_cate',
      #  'num_show_last_day_user_cate', 'num_click_last_day_user_cate',
      #  'num_show_last_hour_user_customer', 'num_click_last_hour_user_customer',
      #  'num_show_last_day_user_customer', 'num_click_last_day_user_customer',
      #  'num_show_last_hour_user_brand', 'num_click_last_hour_user_brand',
      #  'num_show_last_day_user_brand', 'num_click_last_day_user_brand',
      #  'num_show_last_hour_user_all', 'num_click_last_hour_user_all',
      #  'num_show_last_day_user_all', 'num_click_last_day_user_all',
      #  'num_show_last_hour_all_adgroup', 'num_click_last_hour_all_adgroup',
      #  'num_show_last_hour_all_cate', 'num_click_last_hour_all_cate',
      #  'num_show_last_hour_all_customer', 'num_click_last_hour_all_customer',
      #  'num_show_last_hour_all_brand', 'num_click_last_hour_all_brand'
       ]  

categories = ['cate_id','customer', 'brand', 'cms_segid', 'cms_group_id', 'final_gender_code', 'age_level',
       'pvalue_level', 'shopping_level', 'occupation', 'new_user_class_level','hour','weekday']
       
valid_cols = numerics + categories + ['clk']

for col in numerics:
  
  valid_data[col].fillna(0, inplace=True)
  # valid_data[col] = pd.qcut(valid_data[col], 20, labels=False, duplicates = "drop")

import datetime
def df2libffm(df, num_train, save_file=None, y=None, non_categorical_cols = []) :
    
    assert y is not None
 
    row_cnt = df.shape[0]
 
    out = pd.DataFrame({
        y: df[y]
    })
 
    print("out.shape:", out.shape )
    df = df.drop(columns=[y], axis=1)
    
    feature_base = 0
 
    for idx, col in enumerate(df.columns.tolist()) :
        dt = datetime.datetime.now()

        print(str(dt), idx, col)
 
        cur_field_id = idx
        field_series = pd.Series([cur_field_id] * row_cnt).astype(str)
 
        if col in non_categorical_cols :
            # if not categorical feature, do not consider how many different values
            feature_series = pd.Series([feature_base] * row_cnt).astype(str)
            # print(len(feature_series))
            feature_base += 1
            value_series = df[col].astype(str)
            # print(len(value_series))
 
            new_col = field_series + ":" + feature_series + ":" + value_series
            out[str(cur_field_id)] = new_col.values
            
        else :
            # if is categorical feature
            feature_series = df[col].astype('category').values.codes.astype('int32')
            feature_series = feature_series + feature_base
            feature_series = pd.Series(feature_series).astype(str)
 
            feature_base += feature_series.unique().shape[0]
            print("next feature base:", feature_base)
 
            new_col = field_series + ":" + feature_series + ":1" 
            out[str(cur_field_id)] = new_col.values
 
    if save_file:
        file_name = save_file
        print("save file name:", file_name)
        
        out.loc[:int(num_train)].to_csv(file_name + "_train_validate.txt", sep=" ", header=False, index=False)
        
        # 随机划分
        
        temp_train, temp_val = train_test_split(out.loc[:int(num_train)], test_size=0.2, random_state=123)
        
        temp_train.to_csv(file_name + "_train.txt", sep=" ", header=False, index=False)
        temp_val.to_csv(file_name + "_validate.txt", sep=" ", header=False, index=False)
        
        out.loc[int(num_train):].to_csv(file_name + "_test.txt", sep=" ", header=False, index=False)
 
    return out

df2libffm(valid_data[valid_cols], num_train, "fm", y = 'clk', non_categorical_cols = numerics)

# parameter 1: ffm/fm/lr
# parameter 2: learning rate
# parameter 3: lambda（penalty）
# parameter 4: k
# parameter 5: metric —— 'acc', 'prec', 'recall', 'f1', 'auc'
# parameter 6: epoch

!python /content/drive/MyDrive/Taobao/FFM.py "fm" 0.1 0.00000001 12 'auc' 200 '0320'

pd.read_csv("output_test.txt", header = None).to_csv("/content/drive/MyDrive/Taobao/FM_before.csv", index = False, header = False)
