==> 3_ModelEnsemble.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import numpy as np
import pandas as pd
import math
import gc
import os

valid_data = pd.read_csv('/content/drive/MyDrive/Taobao/valid_data_10_percent.csv')

num_train = len(valid_data[valid_data['userid']<=valid_data.loc[int(0.8*len(valid_data))]['userid']])
num_train

test = valid_data.loc[int(num_train):]
y_test = test['clk'].values

y_preds = pd.concat([
                    pd.read_csv('/content/drive/MyDrive/Taobao/deepfm_rus_20.csv',header=None),
                    pd.read_csv('/content/drive/MyDrive/Taobao/wdl_rus_20.csv',header=None),
                    pd.read_csv('/content/drive/MyDrive/Taobao/gbm_focal.csv',header=None),
                    pd.read_csv('/content/drive/MyDrive/Taobao/FFM_after.csv',header=None),
                    pd.read_csv('/content/drive/MyDrive/Taobao/pnn_after.csv',header=None),
                    pd.read_csv('/content/drive/MyDrive/Taobao/FM_after.csv',header=None),
                    ], axis=1)

num = 2
count = 0
for index, col in y_preds.iteritems():
    if (count == num):
      break
    if (count == 0):
      # pred_ans = np.argsort((np.argsort(col)))
      # pred_ans = (num - count) * np.argsort((np.argsort(col)))
      # pred_ans = col
      pred_ans = (num - count) * col
      isFirst = False
    else:
      # pred_ans += np.argsort((np.argsort(col)))
      # pred_ans += (num - count) * np.argsort((np.argsort(col)))
      # pred_ans += col
      pred_ans += (num - count) * col
    count += 1

from sklearn.metrics import log_loss, roc_auc_score
print("test AUC", round(roc_auc_score(y_test, pred_ans), 4))
